Extract shared code from timer
