+++ kernel/comps/time/src/tsc.rs
@@ -6,10 +6,9 @@
 use alloc::sync::Arc;
 use core::sync::atomic::{AtomicU64, Ordering};
 
-use ostd::arch::{
-    read_tsc,
-    timer::{self, TIMER_FREQ},
-    tsc_freq,
+use ostd::{
+    arch::{read_tsc, timer::TIMER_FREQ, tsc_freq},
+    timer,
 };
 use spin::Once;
