+++ ostd/src/lib.rs
@@ -45,6 +45,7 @@ pub mod prelude;
 pub mod smp;
 pub mod sync;
 pub mod task;
+pub mod timer;
 pub mod trap;
 pub mod user;
